fix(lazy): don't try to copy locales to build dir at run-time
Run most of the functionality from the "build:before" hook so that
we don't attempt to do forbidden and unnecessary things at run-time.

Resolves #826
